Added support for parsing inner ICMP embedded packets
This was a PITA.

Etherparse doesn't know how to handle partial headers and
ICMP standard only returns 8 bytes past the IP header
(which includes the s/d ports and TCP sequence number!)
so had to write a custom parser that would be enough to extract
the Connection Key.

Wrote test first - now it passes.  Yay TDD!

Capture is a live capture from an AWS test machine