bumped pcap buffer - no more dropped outgoing packets!
Weird that it happend in the first place.  In theory the default
buffer size is 1M bytes and snaplen is 64k, so if we're capturing
~2k of data per packet (they're not even full size!) we should be
able to capture 500 packets without loss even with no cpu processing.

But bumping to 8M dramatically reduces the outgoing loss but does
not eliminate it so bumping to 64M does eliminate it.  Maybe
reduce snaplen if it's capturing a snaplen worth of data per packet?

<shrug> better now - come back later