Disable the endhost ping block without SACK
It looks like all major stacks (linux, macos, windows, ios) implement
SACK and handle out-of-window ACKs by using SACK so we can uniquely
map a reply to the specific probe that sourced it (yay).  So we can
remove the more error prone way of mapping a reply to a probe which
assumes an idle connection.