MASSIVE refactor: split out libconntrack
In preparation for using libconntrack in the desktop app.

Because libwebserver depends on libconntract and because cargo
(very reasonably!) does not allow circular dependencies, we had
to remove all references to WebServerContext from libconntrack.

Also had to wrap the asyncio pieces of pcap (which are linux only)
with #[cfg(not(windows))].

Tests seem to pass ... <shrug>