Add a clang-format config for protobuf files
Summary:
This is a clang-format config for formatting protobuf files as well C++.
(Figured I might end the C++ config as well). It's based on the Google style
with some tweaks (most notably using 4 spaces for indentation to make it more
similar to rust and python code indentation).

Test Plan: