Add configuration for Prettier code formatter
Summary:
Prettier seems to be the standard formatter for a lot of web related files
(yml, css, js, html, json, ....). So lets use it and add a consistent
config for it.

Happy for suggestion as to whether we want 4 spaces or 2 spaces to be the
default indentation.

Test Plan:
ran prettier in check mode in the root dir (`prettier -c .`). This is
the list of files it would have formatted. Looks reasonable. I think
I excluded all generated ones