Add protobuf for ConnectioMeasurement
This is a first stab at a protobuf spec to persistently store ConnectionMeasurements.

I'm not sure yet if we want to replace the rust native ConnectionMeasurmet, ProbeReport, etc. classes with protobuf ones or keep the native ones and just convert to protobuf for transmit and storage.

For now, I've written the .proto as if we'll keep the rust native classes and therefore I'm only storing the raw probes and none of the summaries.

There are a couple of things I did here that I think might be worth to also apply to the native rust classes:
1) ProbeReportEntry: rename to Probe(?)
2) ProbeReportEntry: have a single class with all the fields as optionals and then have a "plain" enum to specify what the type of the probe was. IMHO all the rust enum variants have the same fields in general but some are absent depending on the type. Also, since we now have both in_timestamp and out_timetamp, I removed the rtt field since it would be redundant now
3) ProbeReportEntry: renamed some fields. IMHO new names are clearer, but YMMV
4) ProbeReportEntry: record the TTL of the ICMP we receive
6) instead of using map<int, something> I use a list of items. Without comments it isn't clear what the int key actually us (process id, probe_id, etc.)