Add start_tracking_time field to `struct Connection` and GUI
Summary:

Record the first time we see a packet for a flow and bubble that
through `struct ConnectionMeasurements` and the web gui.

NOTE: (with a sigh) that a number of the test data files are serialized
`struct Connection`s in JSON, so when we add a mandatory field to
the struct, we have to also manually add it to the test data.  Should
probably at least format the files to make the output clearer.  But
for later.

Test Plan:

Updated test data and 'cargo t' passes.