Copy the OS DNS cache into our cache at startup (#27)
Summary:

Right now, we're not being able to resolve many of the IP addresses
in the flow tracker from our own DNS cache.  This is the first
of two diffs to try to fix that - the OS has its own DNS cache
so copy that into our's first.  We'll see how well that works.

Note: it was a PITA to try to track down the win32 APIs do get this
info programmatically so we're screen scraping from `powershell Get-DnsClientCache`.

Kinda fugly but moving on.

Test Plan:

Two unittests added - seems good