`cargo fmt` wasn't run - missing the hooks in sl :-( (#36)
* Keep expiring DNS cached entriesin an LRU hash

Summary:

The flow trackers doesn't show a lot of hostnames - mostly IPs.
The connections typically live longer than the DNS cache ttl's, so
when they expire, put them in an LRU cache that can be looked up.

Test Plan:
Augment existing test

* WebGui: collapse the DnsEntries to make it pretty

Summary:

In the DnsTracker view, group the entries by hostname
for a cleaner view.

Need (or start with?) a pause refresh button to make sure
that the Details IPs don't disspear on refresh right away

Test Plan:

* Add start_tracking_time field to `struct Connection` and GUI

Summary:

Record the first time we see a packet for a flow and bubble that
through `struct ConnectionMeasurements` and the web gui.

NOTE: (with a sigh) that a number of the test data files are serialized
`struct Connection`s in JSON, so when we add a mandatory field to
the struct, we have to also manually add it to the test data.  Should
probably at least format the files to make the output clearer.  But
for later.

Test Plan:

Updated test data and 'cargo t' passes.

* Silly polish for the DNS Tracker

Summary:

Took 10 seconds to make it just print the address
if there's only one address

Test Plan:

* `cargo fmt` wasn't run - missing the hooks in sl :-(

Summary:
No code change

Test Plan: