Fix to #34!  Handle connection actions on new connection
Summary:

In issue #34, we would get into a loop fighting with devices
on my home network and thrashing the connection manager.  I captured
part of the problem in wireshark and it was a RST loop from the remote
side and with each RST we would create a new connection (!?) in the
connection tracker.

It turns out the logic inside the connection.update_remote_tcp() was correct
in that it was returning the action to close the connection, but when we
called `Connection::update()` from `ConnectionTracker::new_connection()`,
we were IGNORING the resulting action!


Test Plan:

Rewrote the new connection logic to fix this with a test that verifies it
(e.g., test fails without the re-write and passes with it).