Store the DNS name for remote host in `struct Connection`
Summary:

This got a little complicated.  The DnsTracker only keeps the
cached info about connections and could in theory timeout.  Keep a copy
in each `struct Connection` so that can be persisted where-ever it need be,
e.g., to the storage server.

Also plumbed this information through to the webgui

This got complicated because:

1) ConnectionTracker didn't have a tx reference to its own rx queue, so we
 needed to provide it one.
2) Which meant refactoring a bunch of places it was called/used

Test Plan: