DnsTracker tests and list of non-looked up IPs
Summary:

Fixed our dns-parser dependency to support HTTPS queries
and changed our upstream to point to that.

Validated that we're not getting any more unparsed queries.

Still there's a good of IPs we can't lookup from the DNS trace.

Test Plan:

What if I told you this was the test...