Fix connection lookup issues with mapped v4 addresses
Summary:

When listening in the v4+v6 wildcard `::`, the SocketAddr we get is a
mapped v4 address (e.g, ::ffff:127.0.01). This breaks the connection keys
in libconntrack (since the pcap parsing code parses the packets as plain v4).

To solve this: we unmap mapped v4's in the websocket handler


Test Plan:
added unit test
ran on demo.netdebug.com --> observe that probes now work correctly