Add PTR support to DNS Tracker
Summary:

This is not the final commit but it will be the biggest.

This adds the PTR parsing and stores into the dns cache if we
don't have an A/AAAA record (as PTR is lower quality data from
a human's POV).

The final patch will add doing lookups for any IP we don't have a
forward entry for.

Test Plan:

Three new tests