Move OwnedParsedPacket into a Box for perf
Summary:
OwnedParsedPacket is a fairly large struct since many (all) of the
etherparse header struct it contains are fully stack allocated. As such
moving an OwnedParsedPacket is expensive since all the headers need to be
memcpy'ed on every move. Instead we can simply allocate the OwnedParsedPacket
on the heap by putting it in a Box. Now moving it is cheap.

Test Plan:
`./test.sh`