Add RateEstimators to Connection and ConnectionMeasurements
Summary:

Not doing anything with them yet - will display them with the next
commit.

Had to update all of the stored logs for testing with the new fields - sigh.

TODO: decide if we want to count measurement traffic together or separately
from these counters.

Test Plan:

Manual run for now