OwnedParsedPacket: replace pcap_header w/ timestamp and len fields
Summary:
The pcap_header is problematic since it contains a libc::timeval which has
different integer width on different platforms. Instead we are using a
chrono::DateTime<Utc> and a separate length field.

Note, the analyze.rs tests are currently not working since they try to
deserialize logs that contain the old serialized format. We need to fix
that but since we are not actively using the analysis framework yet, I'll
punt on that for now

Test Plan:
cargo test