Show tx/rx bandwidth in flow tracker
Summary:

See screenshot and title, but kinda works as advertised.  One interesting thing,
since we show in Bytes/s accurately for flows that are typically not sending full
time, we effectively assume the instantenous rate of the flow is typical for the
whole second... which makes it look like the total sum of bandwidths is more than
it really is.

That is, if 10 flows send for 10MBps each for 100 ms in sequence but in reality
the total at any one time is 10MBps, we will still show each flow as 10MBps each.
This is fine for display purposes but I think we'll have to be smart for aggregating
the flows as it's not as simple as just summing them up.

Also had to convert `struct RateEstimator` to internally compute in nanos intead of millis
because Duration.num_micros() can return zero and we were getting 'infinite' rate.

Test Plan:

Ran by hand... need a better GUI test system