Works around #119 - don't unwrap on dns_ptr_decode()
Summary:

dns_ptr_decode() paniced with "Didn't end with known domain"
which means the PTR hostname didn't match "in-addr.arpa" or
"ip6.arpa".

No idea input caused it because the log message didn't say
(who wrote this shit!?).

1. Fixed the log message
2. Replace the unwrap() with a match() to find out next time
it happens and do so without a panic()

Test Plan:

Seems like a pain - will wait for it to happen again and then
capture the malformed name.