Render AggregateCounters to graph.js
Summary:

A bit of a monster patch... my bad.

* Changed the DumpAggregateCounters message to only dump ConnectionTracker counters - simplified the code.
* Added logic to track the chart.js object so we can render it fully on first pass and only update
 	the data on the subsequent passes
* FUGLY: just copied the chart.js and utils.js from the webserver gui rather than refactor... TODO

Convert them to the chart.js data json obj format.
Track whether this is an update vs. a new/full render so the screen doesn't flicker.

TODO : normalize the units to some sort of sane Mbps or so

Test Plan:

YOLO screen shot