Trivial: fix local port binding for DNS ptr lookup
Summary:
Let the OS choose the local port for DNS ptr lookup by using port 0 for bind.
On my machine I get an error when trying to bind 53. (It says in-use but could
also be because I'm not root when I run it (my bpf device is accessible by my
regular user).

Test Plan:
ran it. Observe error message are gone