ConnectionTracker: evict connection based on inactivity
Summary:
This changes the way we store/close/evict connections. We now evict connections
purely based on an inativity timer. This allows us to keep connections that are
closed (FINs, RSTs) around to capture any stray packets. It will also help the
GUI with keeping track of short lived connections.

Since we no longer remove a connection when it's closed (per TCP state machine)
we add a helper functions to indicate if a connection has been closed.
A follow up PR will export the flag to the GUI.

Test Plan:
Added new pcap traces and unit tests to test `is_closed()` and the new eviction
logic. Update the test for the FIN teardown handling accordingly.