Consider a conn closed if we see a RST. Also rename 3-way-close -> 4-way-close.
Summary:
Changing the 3/4-way close logic so that it'll also mark a connection closed on
the first RST we see.

Note, technically a RST needs to be within a certain bound of the current seq
no for the TCP stack to actually treat it as a legitimate RST.

Also happy to discard this change if you prefer the current logic. I'm fine
either way

Test Plan:
updated unit test accordingly.