Create a slide to visualize time for bandwidth graph
Summary:

Giving up for a while on the left-shifting graph - added a slider bar
that looks ugly.  Need to fix the real-time aspects of the system
before this will look nice.

See bugs #176 for examples of things we need to fix.

In the mean time, the milliseconds slider jumps around a lot and looks weird
but weirdly (!!) it's smoother during speedtest.net runs.  This would suggest
we're stalling on some other part of the system... but what?

The RTT between the counters request and reply is ~300-400ms on an idle system - wtf!?
Do we need to implement request latency tracking!?

Test Plan:

New tests