Refactor ProcessTracking to query earlier or queue
Summary:

Per discussion with Gregor:

The previous process tracking logic was 'late binding' meaning
that the process information was only associated with each connection
when it was converted into a ConnectionMeasurements struct which
was only triggered by the UI which was both inefficient, bad code,
and meant the StorageServer didn't have complete data.

Restructure this so that we now at new connection time send a query
to the ProcessTracker for the process state, but if they process
tracker doesn't know (there's a race with it's own internal update
loop), then *queue* the request until the next update and async
send the reply back to the Connection tracker.

TODO: add stats and clean up all of the mess in generating
the connection measurements.

Test Plan:

Added two new unit tests