Electron: bundle and stat desktop binary. Hack in old UI
Summary:
Bundle the desktop binary with Electron and start it when Electron
starts.
If `desktop` is running in Electron we don't want it to serve the
static content or wasm package via HTTP (does are directly bundled in
Electron). So I added a flag to disable it (otherwise desktop crashes
if it can't find these directories).

I also temporarily hooked up the old UI so it renders in Electron.
Caveat it that the bandwidth graphs aren't working because I haven't
included chart.js.

Test Plan:
* run desktop (old style) and make sure I didn't break the UI
* run electron in dev mode (`npm start`) ==> loads up UI
* build app bundle (`npm run make`) and start the app bundle ==> works