Rate limit outgoing probes
Summary:
We limit the number of probe-flights we send. We (a) limit per dst-IP and (b)
we limit how often we use a TTL < 4. (for me the 5th hop is the first ECMP hop /
hop with any difference. The first 4 are always the same).

This is just a rough initial rate limiting. Should be a bit smarter about
that in the long run.

Test Plan:
ran desktop and tcpdump and checked how many packets I see with TTLs == 1 and
TTL == 5