Prettier autoformatter: add files to ignore list. Make .js and .ts formatting same
Summary:
Adding some files to the ignore list.
For netdebug_types.ts: I noticed that whenever rust-analyzer and/or cargo
runs it regenerates it and the formatting is different. But maybe we should not
check it in afterall.

Also, I had set the indentation for .js files to 4. Setting it to 2 to make
it consistent with our .ts code.
(Or should we set our .ts code to 4 spaces)? No preference

Test Plan:
prettier -c .