GUI: Dns tab tweaks
Summary:
I happen to stumble across the way MUI handles color palettes and that you
directly referenc these is `sx` propos. So I'm doing this and I matched
the background color of the table header to match the blue
used by MUI for the Navbar highlights.

Also added width (in percent) to the DNS columns. Looks like that solved
most (all?) of the jumping

Test Plan:
screenshot