Add a websocket root for desktop connections (#236)
Basic handling for a server side websocket.  Just does a basic Hello
reply now... BUT with a 'whats my IP' function built in (kinda -it's 
only the address that was used to connect to us).

Added a new message type: TopologyServerToDesktop and DesktopToTopology

Need to add a test for this somehow but might just do it in the next diff.