BucketedTimeSeries: Parameterize on time source. Change AggregateCounters (#249)
This allows us to use BucketedTimeSeries with a `DateTime<Utc>` as the
source of time (which allows us to use the time from pcap header). Also
changing AggregateCounters to use `DateTime<Utc>` as its source. I
didn't want to do this right away but when I tried to keep
AggregateCounters with Instant the `derive(Deserialize)` failed
because it complained that Instant couldn't be deserialized (Looks like
the magic broken down for this).

Note, I temporarily removed the `derive(TypeDef)` since that complained
that `DateTime<Utc>` doesn't implement TypeDef. Will figure out what to
do in the next diff.

Test Plan:
copy-pasted two unit tests and let them use Utc