Replaced storage server with topology server - mostly (#268)
### Summary:

This is a bigger patch and it's not complete.

Remove the references to the storage server from libconntrack and
webserver and desktop and add the topology_client in its place
in preparation to logging connections back to the topology server
(the webserver, once we rename it).

Currently the topology server doesn't actually do anything with the
messages - we need to change that in the next diff.  It's just a
TODO! when messages are passed to it.

 ### Test Plan:

Existing 'cargo t' passes, but it was a PITA - thanks to Gregor for
the help