Nuked all protobuf, storage server, etc. (#275)
### Summary:

Final part of cleanup removing old code

TODO: still need to finalize plumbing the ConnectionMeasurements
from the client to the server.  Just chunking up the diffs for
ease of review.

Most of the work was rewriting parts of the extract-ips and
router-graphs binaries to be okay with getting ConnectionMeasurements
instead of the protobuf structures

I punted on the test_simple_graph() test because its test file
uses the old db schema and I can't regenerate the data unti
the complete refactor is done.

 ### Test Plan:

cargo clean && cargo t