Add Bandwidth tab in the new UI. (#292)
### Summary:
Finally, we have the bandwidth tap in the new UI. Scrolling works too.
All three charts use the same range for their y-axis.

A couple of more things to do on the UI side:
* Nicer tooltips, i.e., show proper units.
* Auto-scale y-axis (to show KBit/s or MBit/s etc.)
* x-axis labels: right now it shows everything in "seconds ago"
* x-range for 1hour chart.
* current layout is hacky so might want to improve that too.

On the rust side: I ended up writing the logic to prepare the chartjs
data in rust and export that directly to the UI. So in hindsight I
properly didn't need `BidirBandwidthHistory` and `BandwidthHistory`
although it might still be useful for storage.

Old UI will be nuked in the next PR.

 ### Test Plan:
 added unit tests for rust logic.
 will upload screenshot

![Screenshot 2023-11-19 at 8.52.15 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/NGRnlVz3aWBiLb2etNoW/c12ce677-6b4f-4b7f-9b5d-b8601bb48947.png)