Initial CongestedLinks Implementation
 ### Summary:

Created a `CongestionSummary` and a `CongestedLink`
structure to share data between desktop and server and
then the server-side logic to colate link and rtt info
to figure out which links are most congested.

Our definition of congestion is the peak latency divided
by the mean latency on the link.

Lots more to do in this space but this is a big PR as it,
so I'm cutting it here.

 ### Test Plan:
Just scratched the surface with an inital test - need a lot more.