Added UI for Congestion Links (#310)
### Summary:

Fixes #309 but might still have data issues.

We're seeing a further hop come back with a lower RTT and this can
be attributed to processing time.  We need to be careful with this data.

Still a little rough but still a fun result

Flag anything 'red' if the peak > 2x mean or 'orange' if 1.5x mean.

Had to fix

 ### Test Plan:

Ran by hand.