Gui <-> Desktop message types: export to TS, tweak name
 ### Summary:
I'm renaming the "Server" part in the GUI <-> Desktop message enums to
Desktop to make it clear we are not talking about the web or topology
server and to match the names we use elsewhere.

I also added support for IpAddr to the typescript-type-def create (one
line change :-)) and pointed our code to a fork of that repo
(I opened a PR to upstream my fix).

Also, TIL: you can specify dependencies in the root / workspace
Cargo.toml file and than easily refer to them from member / child
crates. Using this for the typescript-type-def create. Eventually we
should do this for all our deps.

 ### Test Plan:
 ```
 cargo build
 npm run lint
 npm run make
 ```