Attempt to autogen a release on push
 ### Summary:

See title...

Following instructions from https://github.com/marketplace/actions/automatic-releases

Split the existing logic which was both 'on PR' and 'on push' into two
with a common-build.yml shared logic.  If this is a push, then also
auto-cut a 'latest' release as a github release.

Also moved clippy earlier and added --release so we don't build twice.
Also renamed the files for sanity.  Also added --release to wasm-pack
which is the default but this way it's clearer.

 ### Test Plan:

 Completely untested... that's why I'm pushing this.