Create an connection identifier that's just a plain string (for UI). (#334)
### Summary:
The UI needs to have a unique identifier for a connection. E.g., in the
flow datagrid table to identify the row. Eventually, we'll also need a
way for the UI to request more information for a particular connection
(or group of connections). So having a string key for that is useful
too.

So, we create a small opaque string representation for ConnectionKey.
Questions:
* why not `to_string()`? the `to_string()` output is harder to parse
* why wrap it in `ConnectionIdString`? To get strong typing. One can't
  put a random string in there by accident.

 ### Test Plan:
 create a new unit test.