Disable probe rate limiting for the webserver
 ### Summary:

The probing rate limits (effectively in 1.5 and 2 seconds) are
completely wrong for the webserver and explains why the webserver
probing was broken.

After substantial discussion, I've added a unlimited_probes flag
to the ConnectionTracker::new() constructor
which bypasses (admittedly, in a hacky way) all of the rate limiting
and my manual tests for the webserver now pass.

TODO: probably ConnectionTracker constructor needs to be a builder

This PR adds a bunch of debug!() statements that I think might be
useful in the future so I've left them there but feedback is welcome.

ALSO refactored a lot of the ConnectionTracker tests with a helper
function to make a connectiontracker - much less duplicate code!

 ### Test Plan:

Still need some form of end-to-end tests/monitoring :-(

cargo t passes