Add TLS support by default for webserver in --production mode
 ### Summary:

See title.

If you set --production, we load a TLS Server by default.
If you don't set --production, you get an HTTP/unencrypted server

Add the following flags to make this happen:
--tls-cert /path/to/cert.pem
--tls-key /path/to/key.pem
--force-unencypted     # override the production = tls logic

All of the defaults assume a setup as created by Certbot

 ### Test Plan:

Seems hard to test?

Manually started it in prod on my machine and it correctly complained
that 'localhost' was not 'topology.netdebug.com'