Convert SystemTracker to a shared lock (#427)
### Summary:

Most of the SystemTracker's state does not lend itself to the mesasge
passing pattern and creates a lot of unececssary boiler plate code.

This diff converts it from the message passing tracker scheme to just a
shared lock.  All of the tests seem to pass including a manual
end-to-end test where I started the electron app, turned off my network
and turned itback on and we got the right log messages and counters
incremented.

I tried and failed to convert SystemTracker::new() to be sync :: the
fact that the get_default_gateways() code depends on a async 3rd party
library means we need to go from async to sync back to async which
causes
problems as outlined in

https://stackoverflow.com/questions/62536566/how-can-i-create-a-tokio-runtime-inside-another-tokio-runtime-without-getting-th

I lost a bunch of time trying to get their suggested solution to work
(spawn a new OS-level thread) but ultimately decided I was wasting too
much time and kept the whole SystemTracker  async.

[disagree and commit]
 ### Test Plan:

cargo t

and manually changing my local network while electron was running