Track TCP timestamps and RTT. (#430)
### Summary:
Track TCP timestamps and use them to compute an RTT estimate. Right now we just use a "normal" exponential decay. In future PRs we'll track the RTT history a bit smarter and with more bells and whistles. 

The principle here is that for packets that are sent by a side, we keep a history packet_timestamp, tsval pairs. When we receive an ACK, we check the history to see if any timestamp value matches tsecr. If so, we compute the time difference using the *packet_timestamp*. I.e., we treat tsvals as opaque, monotonically increasing values. 

 ### Test Plan:
added unit tests. 
Will add a test with a trace in a follow up PR.