Trying to properly handle pkt len with TSO
 ### Summary:
When Tcp Segment Offloading is enabled on a machine,
with ip6, the hardware computes the ip6 total_len field, but only
after pcap has captured it.  This means when we see it, it could be
zero (!!).  This diff tries to hack around this by saying
"the payload len is at least the size of what we captured" but we could
be more aggressive and decide it's based off of OwnedParsedPacket.len

Need to understand what assumptions Gregor is making with this number in
the sequence analysis.

 ### Test Plan:
existing 'cargo t' - should I add a plen=0 test?