Add 'vendor_oui' field to ExportedNeighborState (#457)
### Summary:
Use this to fill in the 'vendor' field in the UI for teh devices page.

I found a rust crate 'mac_oui' (which sounds like a French Apple
fanatic) that downloads from an online DB one time at startup and gives
all of the relevant info.  Seems to work
 ### Test Plan:

See screenshot