Hack Mac looksups for IPv6 link local addresses (#496)
### Summary:
A link local address has the mac address encoded into it
using "EUI-64" (my thoughts on the quality of that are
in the comments) and it seems like my home router only
sometimes will send an Neighbor Advertisement for it,
which is fair because we shouldn't be sending Neighbor
Discovery messages for a MAC we could infer ourselves.

Hack it into the inferernce loop
 ### Test Plan:
Added some EUI-64 tests.  What a PITA