Change desktop build command to be more precise (#499)
### Summary:
Back when we needed wasm in the build, it made sense to use
the ./build.sh script to build all of the binaries and the wasm
for use with the electron GUI.  But now, we're only using the
desktop binary without WASM and we've added a bunch of other
binaries (webserver, probe-o-matic, etc.) which all take more
time to build.  Changed to only build the desktop binary.

 ### Test Plan:
Manual