Give default route a hint for which gw to ping
 ### Summary:
See #520 but we can get the wrong gw to ping
when we have multiple interfaces, specifically on
windows where when an interface goes down/away/gets unplugged,
the route table does not update and those old routes still
show up as preferred by the route metric (must be some other
logic somewhere that ignores them).

Use the following logic to figure out which gateways to ping:

1) If there's an fe80:: link-local address on the egress interface,
then use a trick to turn that into an interface_index and only
use routes that have the same interface_index.
2) If there's not a fe80 link-local address, use the
broadcast address with the egress interface to disambiguate which
gw address we need.
3) If there's none of those, don't prune down the gateways and just
 show all of them to the UI <shrug>
 ### Test Plan:

 Hard to automatically test weird OS interactions with the routing
 table.  Need to eventually us the 'tun' crate.  But for now,
 manual testing with the GUI and unplugging my cable and replugging
 it shows this mostly working (but other bugs).