Pick the link-local src addr more carefully on ping
 ### Summary:
Link-local IP addresses are a PITA because they break
our "use the route table to figure out the source ip"
trick.  They break this trick because every v6 capable
interface has a valid and *equal cost* fe80:: route, so
we just get an arbitrary one which is often not correct.

This code for pinging changes this logic to use the link-local
address of the egress interface for the ping's src-ip.
 ### Test Plan:
Existing tests plus manual