Fixed 563: Explicit sort function for rtt data (#567)
### Summary:
See issue, but Javascript's default sort() function
on array is "convert to string and sort by ASCII" so
sorting an array of numbers will result in 100 coming before 3.

WTF...
 ### Test Plan:

The only thing being tested here is my patience... at the same time,
I've added a check in the console logs to see if we see this again.