Fixed rtt boxplot on home page (#571)
### Summary:
The boxplot wasn't updating across network interface changes.
I moved it to use the nicer renderDataLoadingState primitive
that Gregor put together but that didn't solve the problem.

The deeper problem was that we were rendering the first network
history but the latest is actually the LAST (durh), so all of the
right state was showing up across interface changes, but we were
only rendering the oldest one so saw no change.
 ### Test Plan:
Manual testing with taking cable out and turning off wifi.  Did
it lots and seemed to keep up...