Better handle no pings/no default route cases (#572)
### Summary:
Gregor pointed out that if we have no pings, we show 0% packet loss.
Instead show 100% packet loss.

Also if we have no default gateways, show a label in the middle of the
graph saying "No IPv4 default routers found!".
 ### Test Plan:
The 100% packet loss is easy to test; just unplug/turn off all networks.

The no gateways is hard to test as windows keeps the gateways even
when the interfaces are unavailable so I might need Gregor to test that.