Make prettier happy? (#582)
### Summary:
Looks like my local prettier version and the one on GH disagreed about some formatting again. To make it easier to debug and prevent this, I'm adding an action to print the current prettier version in the GH run and I'm pinning the version for the local commit hook. 

TODO: not sure what version vscode uses for auto-formatting. Will cross that bridge when it starts failing. 

 ### Test Plan:
yolo