Don't assume IPv6 Link-Local to MAC mapping
 ### Summary:
Wow this stuff is complicated.  Previously, we
assumed that if a router advertised a Link-Local address
that it would accept traffic to the corresponding MAC.

This apparently isn't true and Gregor's home router doesn't
do this (which I still think is bogus).  So change the logic
to send a NDP Neighbor Discovery message for Link-Local traffic
to get the right mac.

This code change means that we needed to better propapgate our
"if the IP is link-local, don't use the routing table to infer the
source-address" trick, so I've moved it into remote_to_local_ip().

Also, I found out that our NDP Discovery messages were not RFC compliant
and that's why my router was ignoring them.  They need to
1) have hop-limit=255
2) Encode the last 3 bytes of target addr in dst_mac and dst_ip
3) Have the Source-link-layer option

As a result of that change, I've changed the return type of
remote_to_local_ip() to pcap::Error as it simplifies things.

 ### Test Plan:

Lots of manual testing.