Possible fix to 'No v4 Gateways' rendering bug
 ### Summary:
It seems that 'annotations' (the way we're adding that text)
are a bit of a mess in ApexCharts.  If we were manually managing
our chart directly (instead of via react), then we could call
chart.clearAnnotations().  We were passing "no annotations" but
it wasn't clearing them which is probably a bug.  I can't repro
the problem on either my windows or MAC but being more specific
"no yaxis" annotations might trigger the right thing in react?

 ### Test Plan:
verified it doesn't make anything worse...