Set 'tokio_unstable' as default for whole workspace (#618)
### Summary:

'tokio_unstable' is a compile-time flag (read: set at
.cargo/config.toml, not a feature flag) that enables a
bunch of runtime metrics.  From reading :

https://docs.rs/tokio/latest/tokio/#unstable-features

It is the *APIs* not the code that is unstable, so this
seems reasonable.

This PR is just the flag enabling and the workspace code
(plus a minor test to output the number of workers to make sure
it's working)

TODO: actually enable the tokio rust console from
https://github.com/tokio-rs/console

 ### Test Plan:

*** note that "with 16 workers" which we can only get with metrics that
are only available when tokio_unstable is set.

[robs@LAPTOP-O7RS71IJ netdebug]$ sudo ./target/debug/netdebug-desktop
2024-02-20T18:09:43.920Z INFO netdebug_desktop L82 > Current tokio scheduler flavor is: MultiThread with 16 workers
2024-02-20T18:09:43.923Z INFO netdebug_desktop L188 > Running desktop version: 4fdb011d947dca49d11dda6b1233b20b684ccc61
2024-02-20T18:09:43.923Z INFO netdebug_desktop L194 > Starting Axum